- Paged Table height based on pageSize
This commit calculates the height of the data grid based on the
pageSize variable. Consider this as the first step of changing the
amount of items per page which will change the height of the data grid
accordingly.
Notice that before this all the grids were 300px height, unless it was
override.