Fix bug in old optimistic locking logic (see below)
Calling init() reloads the latest version of the content and therefore
needs to reset the concurrent update state. There is no reason to show
the concurrent modification popup if they user worked off the latest
version (when it was reloaded after they opened the editor).

We now have a new pessmistic locking approach in UberFire but want to
keep this one as a fallback.